feat: movendo types do grid para /types
