+++ apps/medicamentar-desktop/src/types/actionTypes.ts
@@ -1,0 +1,5 @@
+export type actionTypes = 
+    "warning" |
+    "alarmAnswered" |
+    "deleted" |
+    "created";
