+++ apps/medicamentar-desktop/src/types/gridItemTypes.ts
@@ -1,0 +1,7 @@
+export type gridItemTypes = {
+    title: string,
+    actionType: string,
+    action: string,
+    date: string,
+    medic?: string
+};
